main_svc1

+++ ewm-stats/stat-dto/src/main/java/ru/practicum/ViewStatsDto.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class ViewStatsDto {
     private String app;
     private String uri;
-    private Integer hist;
+    private Long hist;
 }
